requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pymongo
bcrypt
python-dotenv
python-multipart
